# opt: try to reduce key in FuncDepSet.AddFrom