--- a/pkg/sql/opt/memo/testdata/logprops/set
+++ b/pkg/sql/opt/memo/testdata/logprops/set
@@ -10,6 +10,10 @@ exec-ddl
 CREATE TABLE abc (a INT PRIMARY KEY, b INT, c INT)
 ----
 
+exec-ddl
+CREATE TABLE mn (m INT, n INT, PRIMARY KEY (m, n))
+----
+
 build
 SELECT * FROM xy UNION SELECT * FROM uv
 ----
@@ -639,3 +643,29 @@ except-all
       ├── prune: (2)
       └── tuple [type=tuple{int}]
            └── const: 1 [type=int]
+
+# Regression test for #113817. The key from the LHS of the intersection should
+# be reduced after adding the FDs from the RHS.
+norm
+SELECT m, n FROM mn
+INTERSECT
+SELECT x, y FROM xy
+----
+intersect-all
+ ├── columns: m:1(int!null) n:2(int)
+ ├── left columns: m:1(int!null) n:2(int)
+ ├── right columns: x:5(int) y:6(int)
+ ├── key: (1)
+ ├── fd: (1)-->(2)
+ ├── interesting orderings: (+1,+2)
+ ├── scan mn
+ │    ├── columns: m:1(int!null) n:2(int!null)
+ │    ├── key: (1,2)
+ │    ├── prune: (1,2)
+ │    └── interesting orderings: (+1,+2)
+ └── scan xy
+      ├── columns: x:5(int!null) y:6(int)
+      ├── key: (5)
+      ├── fd: (5)-->(6)
+      ├── prune: (5,6)
+      └── interesting orderings: (+5)