Previously, `FuncDepSet.AddFrom` could add dependencies to an FD without
reducing the key afterward. This could produce FDs with extra key
columns. For example, consider the two FDs below:

    fd1: key(1,2)
    fd2: key(1); (1)-->(2)

Adding `fd2` to `fd1` without reducing the key would produce
`key(1,2); (1)-->(2)` instead of `key(1); (1)-->(2)`. The unreduced key
is not incorrect, but it could prevent some possible optimizations from
triggering. In test builds the unreduced key causes an assertion
failure.

This commit fixes the issue by calling `FuncDepSet.tryToReduceKey` after
dependencies have been added within `FuncDepSet.AddFrom`. This fixes an
assertion failure in test builds when optimizing some queries with
`INTERSECT` expressions.

A test has also been added for a similar case with
`FuncDepSet.AddEquivFrom`, but no code changes are required because this
function calls `FuncDepSet.addEquivalency` which already reduces the
FD's key.

Fixes #113817

There is no release note because the assertion failure is only present
in test builds and the issue does not cause correctness bugs in release
builds.

Release note: None