@mgartner 2023-11-13
Looks like the performance hit is minimal:


```
name                         old time/op    new time/op    delta
SlowQueries/slow-query-1-24    28.0ms ± 1%    27.3ms ± 1%  -2.49%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24     417ms ± 1%     416ms ± 1%    ~     (p=1.000 n=5+5)
SlowQueries/slow-query-3-24    96.3ms ± 0%    96.2ms ± 1%    ~     (p=0.841 n=5+5)
SlowQueries/slow-query-4-24     687ms ± 1%     680ms ± 0%  -0.99%  (p=0.016 n=5+5)
SlowQueries/slow-query-5-24     1.52s ± 1%     1.53s ± 1%    ~     (p=0.151 n=5+5)

name                         old alloc/op   new alloc/op   delta
SlowQueries/slow-query-1-24    6.81MB ± 0%    6.81MB ± 0%    ~     (p=0.421 n=5+5)
SlowQueries/slow-query-2-24    48.5MB ± 0%    48.5MB ± 0%    ~     (p=0.841 n=5+5)
SlowQueries/slow-query-3-24    44.6MB ± 0%    44.7MB ± 0%  +0.12%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-24     153MB ± 0%     153MB ± 0%  +0.11%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     290MB ± 0%     291MB ± 0%  +0.29%  (p=0.029 n=4+4)

name                         old allocs/op  new allocs/op  delta
SlowQueries/slow-query-1-24     69.6k ± 0%     69.7k ± 0%  +0.01%  (p=0.008 n=5+5)
SlowQueries/slow-query-2-24      391k ± 0%      391k ± 0%    ~     (p=0.690 n=5+5)
SlowQueries/slow-query-3-24      385k ± 0%      387k ± 0%  +0.44%  (p=0.008 n=5+5)
SlowQueries/slow-query-4-24     3.22M ± 0%     3.23M ± 0%  +0.15%  (p=0.008 n=5+5)
SlowQueries/slow-query-5-24     5.34M ± 0%     5.37M ± 0%  +0.49%  (p=0.008 n=5+5)
```

```
name                                                    old time/op    new time/op    delta
Phases/kv-read/Simple/Explore-24                          50.3µs ± 1%    49.8µs ± 0%  -1.04%  (p=0.016 n=5+5)
Phases/kv-read-const/Simple/Explore-24                    50.5µs ± 1%    50.2µs ± 1%    ~     (p=0.421 n=5+5)
Phases/tpcc-new-order/Simple/Explore-24                    134µs ± 1%     135µs ± 0%    ~     (p=0.056 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                     114µs ± 1%     114µs ± 0%    ~     (p=0.421 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-24                  43.0µs ± 0%    43.1µs ± 1%    ~     (p=0.690 n=5+5)
Phases/tpcc-stock-level/Simple/Explore-24                  598µs ± 0%     598µs ± 1%    ~     (p=0.841 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24                259µs ± 0%     261µs ± 1%  +0.74%  (p=0.016 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24        467µs ± 0%     467µs ± 0%    ~     (p=0.548 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24      414µs ± 0%     414µs ± 0%    ~     (p=1.000 n=5+5)
Phases/many-columns-and-indexes-b/Simple/Explore-24        554µs ± 0%     556µs ± 0%    ~     (p=0.151 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24      468µs ± 1%     470µs ± 1%    ~     (p=0.222 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        303ms ± 3%     307ms ± 2%    ~     (p=0.222 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      307ms ± 1%     304ms ± 1%    ~     (p=0.095 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24        138µs ± 1%     136µs ± 1%  -1.27%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-24     58.5µs ± 0%    58.3µs ± 0%    ~     (p=0.095 n=5+5)
Phases/batch-insert-one/Simple/Explore-24                  120µs ± 1%     118µs ± 1%  -1.53%  (p=0.008 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                8.76ms ± 1%    8.74ms ± 0%    ~     (p=0.222 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                   3.70ms ± 0%    3.68ms ± 0%  -0.47%  (p=0.008 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24      3.48ms ± 0%    3.48ms ± 0%    ~     (p=0.421 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24    1.53ms ± 0%    1.52ms ± 0%  -0.44%  (p=0.016 n=5+5)

name                                                    old alloc/op   new alloc/op   delta
Phases/kv-read/Simple/Explore-24                          15.3kB ± 0%    15.3kB ± 0%    ~     (all equal)
Phases/kv-read-const/Simple/Explore-24                    15.4kB ± 0%    15.4kB ± 0%    ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                   32.7kB ± 0%    32.7kB ± 0%    ~     (p=0.706 n=5+5)
Phases/tpcc-delivery/Simple/Explore-24                    26.7kB ± 0%    26.7kB ± 0%    ~     (p=0.357 n=5+5)
Phases/tpcc-delivery/Prepared/Explore-24                  12.8kB ± 0%    12.8kB ± 0%    ~     (p=0.968 n=5+4)
Phases/tpcc-stock-level/Simple/Explore-24                  130kB ± 0%     130kB ± 0%    ~     (p=0.548 n=5+5)
Phases/tpcc-stock-level/Prepared/Explore-24               58.9kB ± 0%    58.9kB ± 0%    ~     (p=1.000 n=5+5)
Phases/many-columns-and-indexes-a/Simple/Explore-24       39.6kB ± 0%    39.6kB ± 0%    ~     (p=0.460 n=5+5)
Phases/many-columns-and-indexes-a/Prepared/Explore-24     5.50kB ± 0%    5.50kB ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24       47.9kB ± 0%    47.9kB ± 0%    ~     (p=0.063 n=5+5)
Phases/many-columns-and-indexes-b/Prepared/Explore-24     8.02kB ± 0%    8.02kB ± 0%    ~     (p=0.365 n=5+5)
Phases/many-columns-and-indexes-c/Simple/Explore-24        109MB ± 0%     109MB ± 0%  +0.05%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24      108MB ± 0%     108MB ± 0%  +0.05%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24       29.7kB ± 0%    29.7kB ± 0%    ~     (p=0.587 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-24     13.0kB ± 0%    13.0kB ± 0%    ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                 64.8kB ± 0%    64.8kB ± 0%    ~     (p=0.397 n=5+5)
Phases/batch-insert-many/Simple/Explore-24                3.92MB ± 0%    3.92MB ± 0%    ~     (p=0.548 n=5+5)
Phases/ored-preds-100/Simple/Explore-24                    945kB ± 0%     945kB ± 0%    ~     (p=0.889 n=5+5)
Phases/ored-preds-using-params-100/Simple/Explore-24       926kB ± 0%     926kB ± 0%    ~     (p=0.548 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     736kB ± 0%     736kB ± 0%    ~     (p=0.389 n=5+5)

name                                                    old allocs/op  new allocs/op  delta
Phases/kv-read/Simple/Explore-24                            95.0 ± 0%      95.0 ± 0%    ~     (all equal)
Phases/kv-read-const/Simple/Explore-24                      95.0 ± 0%      95.0 ± 0%    ~     (all equal)
Phases/tpcc-new-order/Simple/Explore-24                      175 ± 0%       175 ± 0%    ~     (all equal)
Phases/tpcc-delivery/Simple/Explore-24                       165 ± 0%       165 ± 0%    ~     (all equal)
Phases/tpcc-delivery/Prepared/Explore-24                    75.0 ± 0%      75.0 ± 0%    ~     (all equal)
Phases/tpcc-stock-level/Simple/Explore-24                    708 ± 0%       708 ± 0%    ~     (all equal)
Phases/tpcc-stock-level/Prepared/Explore-24                  341 ± 0%       341 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-a/Simple/Explore-24         81.0 ± 0%      81.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-a/Prepared/Explore-24       31.0 ± 0%      31.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Simple/Explore-24          136 ± 0%       136 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-b/Prepared/Explore-24       49.0 ± 0%      49.0 ± 0%    ~     (all equal)
Phases/many-columns-and-indexes-c/Simple/Explore-24         729k ± 0%      731k ± 0%  +0.25%  (p=0.008 n=5+5)
Phases/many-columns-and-indexes-c/Prepared/Explore-24       724k ± 0%      725k ± 0%  +0.24%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-24          522 ± 0%       522 ± 0%    ~     (all equal)
Phases/single-col-histogram-range/Prepared/Explore-24        251 ± 0%       251 ± 0%    ~     (all equal)
Phases/batch-insert-one/Simple/Explore-24                    325 ± 0%       325 ± 0%    ~     (all equal)
Phases/batch-insert-many/Simple/Explore-24                 28.5k ± 0%     28.5k ± 0%    ~     (p=1.000 n=5+4)
Phases/ored-preds-100/Simple/Explore-24                    4.53k ± 0%     4.53k ± 0%    ~     (p=0.238 n=4+5)
Phases/ored-preds-using-params-100/Simple/Explore-24       4.53k ± 0%     4.53k ± 0%    ~     (p=0.444 n=5+5)
Phases/ored-preds-using-params-100/Prepared/Explore-24     2.88k ± 0%     2.88k ± 0%    ~     (all equal)
```